# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SmartOrder AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /business:
    put:
      operationId: updateBusiness
      tags: [business]
      summary: Update business settings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /business/{slug}:
    get:
      operationId: getBusinessBySlug
      tags: [business]
      summary: Get public business info and products
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicBusiness"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Products
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients with search and sort
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client profile with order history
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Campaigns
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

  # Transactions
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Auth
...
    Business:
      type: object
      required: [id, slug, name, type, email, phone, plan, createdAt]
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        type:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        plan:
          type: string
        createdAt:
          type: string
          format: date-time
    BusinessUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
    PublicBusiness:
      type: object
      required: [id, slug, name, type, phone, products]
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        type:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"

    # Product
    Product:
      type: object
      required: [id, businessId, name, price, category, available, stockQty, createdAt]
      properties:
        id:
          type: string
        businessId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        category:
          type: string
        photoUrl:
          type: ["string", "null"]
        available:
          type: boolean
        stockQty:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    OrderItem:
      type: object
      required: [productId, name, price, quantity]
      properties:
        productId:
          type: string
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
    Order:
      type: object
      required: [id, businessId, clientName, clientPhone, items, total, depositAmount, status, orderType, createdAt]
      properties:
        id:
          type: string
        businessId:
          type: string
        clientName:
          type: string
        clientPhone:
          type: string
        clientEmail:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        depositAmount:
          type: number
        status:
          type: string
        orderType:
          type: string
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Client:
      type: object
      required: [id, businessId, name, phone, totalOrders, totalSpent, createdAt]
      properties:
        id:
          type: string
        businessId:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        totalOrders:
          type: integer
        totalSpent:
          type: number
        lastOrderAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ClientProfile:
      type: object
      required: [id, businessId, name, phone, totalOrders, totalSpent, createdAt, orders]
      properties:
        id:
          type: string
        businessId:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        totalOrders:
          type: integer
        totalSpent:
          type: number
        lastOrderAt:
          type: ["string", "null"]
          format: date-time
        favoriteProducts:
          type: ["object", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"

    # Campaign
    Campaign:
      type: object
      required: [id, businessId, name, segment, channel, message, status, sentCount, createdAt]
      properties:
        id:
          type: string
        businessId:
          type: string
        name:
          type: string
        segment:
          type: string
        channel:
          type: string
        message:
          type: string
        status:
          type: string
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        sentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CampaignInput:
      type: object
      required: [name, segment, channel, message]
      properties:
        name:
          type: string
        segment:
          type: string
        channel:
          type: string
        message:
          type: string
        scheduledAt:
          type: string

    # Transaction
...