# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartOrder AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: string;
  businessId: string;
  clientName: string;
  clientPhone: string;
  /** @nullable */
  clientEmail?: string | null;
  items: OrderItem[];
  total: number;
  depositAmount: number;
  status: string;
  orderType: string;
  /** @nullable */
  scheduledAt?: Date | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}