# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartOrder AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  productId: string;
  name: string;
  price: number;
  quantity: number;
}